Move consent state methods into consent_state mod
